refactor : signinScene 클래스/헤더/ui 파일 이름 수정 및 클래스 명 수정
추가로 connect 함수에 QObject 위치 지정자 부분도 지웠습니다.
